backend/requirements.txt
django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary>=2.9.9
Pillow>=10.4.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
